feat: create metadata and put it icon page
