+++ app/layout.tsx
@@ -1,70 +1,18 @@
-'use client';
-import React, { useEffect, useState } from 'react';
-import { ThemeProvider, CssBaseline, Box } from '@mui/material';
-import { createAppTheme } from './utils/theme';
-import Header from './components/Header';
-import LoadingScreen from './components/LoadingScreen';
-import { AnimatePresence } from 'framer-motion';
-import { lexendFont, valorantFont } from './utils/fonts';
+import type { Metadata } from 'next'
+import RootLayoutClient from './RootLayoutClient'
+
+export const metadata: Metadata = {
+  title: 'GYCODING',
+  description: 'Desarrollo web profesional',
+  icons: {
+    icon: '/gy-icon-small.ico',
+  },
+}
 
 export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const [mode, setMode] = useState<'light' | 'dark'>('light');
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const initializeApp = async () => {
-      const minLoadTime = new Promise(resolve => setTimeout(resolve, 1000));
-
-      const savedTheme = localStorage.getItem('theme') ||
-        (window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light');
-
-      setMode(savedTheme as 'light' | 'dark');
-
-      await minLoadTime;
-
-      setIsLoading(false);
-    };
-
-    initializeApp();
-  }, []);
-
-  const toggleTheme = () => {
-    const newMode = mode === 'light' ? 'dark' : 'light';
-    setMode(newMode);
-    localStorage.setItem('theme', newMode);
-  };
-
-  const theme = createAppTheme(mode);
-
-  return (
-    <html lang="es" className={`${lexendFont.className} ${valorantFont.className}`}>
-      <body>
-        <ThemeProvider theme={theme}>
-          <CssBaseline />
-          <AnimatePresence mode="wait">
-            {isLoading ? (
-              <LoadingScreen key="loading" />
-            ) : (
-              <Box
-                key="content"
-                sx={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  minHeight: '100vh',
-                  width: '100%'
-                }}
-              >
-                <Header onThemeToggle={toggleTheme} />
-                {children}
-              </Box>
-            )}
-          </AnimatePresence>
-        </ThemeProvider>
-      </body>
-    </html>
-  );
+  return <RootLayoutClient>{children}</RootLayoutClient>;
 }
